Staff deletion function created. doDelete() servlet method, delete() service class method, delete() repo method created and added functionalities.

+++ src/main/java/com/pahanaedu/common/interfaces/IRepositoryPrototype.java
@@ -7,6 +7,6 @@ public interface IRepositoryPrototype<T, R extends T> {
     R findById(Long id);
     List<R> findAll();
     R save(R obj);
-    boolean delete(int id);
+    boolean delete(Long id);
 
 }
